HACK: Handle new UUIDs in SwitchBot firmware
The new SwitchBot firmware advertises its services differently, making
the old UUID filter not find any SwitchBot devices. This commit is a
quick-fix to remove the filter and simply allow the rest of the
discovery code to handle it as before.

This should be backwards compatible with the old firmware since the rest
of the discovery and initialization is exactly the same.

See also https://github.com/devWaves/SwitchBot-MQTT-BLE-ESP32/issues/80